Correct versions for Firefox for Window API (P1) (#7533)
This commit is part 1 of a few to correct the versions for Firefox for
various features of the Window API based upon results from the
mdn-bcd-collector project.  This commit specifically touches data
where the version added is reduced, or the data is changed from
`false` to a version added+removed range.