Remove incorrect claims of prefixed Web Speech interfaces
The set of webkitSpeech* members on the global object are:
 - webkitSpeechGrammar
 - webkitSpeechGrammarList
 - webkitSpeechRecognition
 - webkitSpeechRecognitionError
 - webkitSpeechRecognitionEvent

These 3 other SpeechRecognition* interfaces are instead decorated with
[LegacyNoInterfaceObject] in Chromium source, which means their
interfaces objects are not exposed at all. They don't have constructors,
as they are accessed from the `SpeechRecognitionResultList` at
`event.results` for a "result" event.

So while use of prefixes is required to get to these APIs, they are
not themselves prefixed.

See also https://github.com/mdn/browser-compat-data/pull/7554.