Edge 87 adds support for text-[decoration-thickness,underline-offset]
Also add support for `from-font` in `text-underline-position`.

The same changes that were made for Chrome 87 in 116a3bd also apply to Edge 87.

Verified by testing in Edge 86 and 87 on BrowserStack, with the CSS:

```
a {
  text-decoration-color: red;
  text-decoration-thickness: 3px;
  text-underline-offset: 3px;
}
```

In Edge 86, the underline is red but only the default user agent thickness, and not offset. In Edge 87, it has 3px thickness and is offset, and has the same appearance that it does in Chrome 87)