Note issue w/ text-decoration-thickness in Chrome, Edge
The `text-decoration-thickness` property does not work unless either `text-underline-offset` is set to something other than `auto` or `text-decoration-color` is set to something other than `currentColor`.

See https://crbug.com/1154537.

Still seeing this in Chrome 88 / Edge 88 betas, but Chrome Canary 89 does not have this issue. However, I’m not sure exactly how the Chrome / Edge release process works and so am not sure if this fix might end up in Chrome 88 as well, so not able to provide a bounded version range until Chrome 88 is stable.