Add missing XR APIs and API features (#7645)
This PR is a part of a project to add missing interfaces and interface features to BCD that are from a spec in Editor's Draft or more and is supported in at least one browser.  This particular PR adds missing features, populating the results using data from the [mdn-bcd-collector](https://mdn-bcd-collector.appspot.com) project (v1.1.6), for the XR APIs.

Specs:
https://immersive-web.github.io/anchors/
https://immersive-web.github.io/dom-overlays/
https://immersive-web.github.io/hit-test/
https://immersive-web.github.io/layers/
https://immersive-web.github.io/webxr/
https://immersive-web.github.io/webxr-ar-module/
https://immersive-web.github.io/webxr-hand-input/

Mark all entries as experimental.