Add ClientRect variant of DOMRect for Safari
This was added with the implementation of Element.getBoundingClientRect
and Element.getClientRects:
https://trac.webkit.org/changeset/40837/webkit

Also checked but already correct was the addition of
Range.getClientRects/getBoundingClientRect, which came slightly later:
https://trac.webkit.org/changeset/48806/webkit

For the removal date, the transition to DOMRect for these APIs was used:
https://trac.webkit.org/changeset/215892/webkit
https://trac.webkit.org/changeset/215946/webkit

Indeed DOMRect was added some time before this:
https://trac.webkit.org/changeset/207438/webkit

The Safari versions are derived using the method in
https://developer.mozilla.org/en-US/docs/MDN/Contribute/Processes/Matching_features_to_browser_version#safari

Finally, http://software.hixie.ch/utilities/js/live-dom-viewer/?saved=8797
was tested in Safari 4 to verify it was supported then. It couldn't have
been in Safari 3.2 since that was released before the change itself.

Similar to https://github.com/mdn/browser-compat-data/pull/8711.