Add ClientRect variant of DOMRect for Chromium browsers
ClientRect was replaced by DOMRect in Chromium 61:
https://storage.googleapis.com/chromium-find-releases-static/f01.html#f01e385c3f752c2bfd1163db74ecbf88e5390d25

The addition of ClientRect is more complicated, as it goes back to
WebKit, Presto and Trident.

WebKit: https://trac.webkit.org/changeset/40837/webkit

WebKit trunk then had version 530.0.0:
https://trac.webkit.org/browser/webkit/trunk/WebCore/Configurations/Version.xcconfig?rev=40837

The versions for Chrome, Samsung Internet and WebView are
based on that WebKit version.

For Trident/Presto, the data for api.Element.getBoundingClientRect is
trusted, giving IE 4 and Opera 9.5 as the version_added.

The getBoundingClientRect version for WebView is also updated to match,
where WebView cannot be right because it was released before the change
to WebKit trunk.

Similar to https://github.com/mdn/browser-compat-data/pull/8714.