Add the original HTMLMediaError alias for Firefox (#8734)
MediaError is the type of object used for video.error, which was
shipped in Firefox 3.5. When TimeRanges was being implemented for
Firefox 4, it was discovered that MediaError was incorrectly called
HTMLMediaError, and manual testing in Firefox 3.6 confirms that
interface object was exposed. It was fixed here:
https://bugzilla.mozilla.org/show_bug.cgi?id=589561#c4
https://hg.mozilla.org/mozilla-central/rev/37f61e9d618a

The point of this is to avoid the situation in
https://github.com/mdn/browser-compat-data/pull/7963, where looking for
MediaError makes it seem like it was shipped in Firefox 4.