Drop reference to wpt.fyi results for api.Animation.commitStyles (#7233)
The test results are for Safari on macOS. But more importantly, the
existence of failing tests doesn't in itself indicate that there's a
problem that matters to web developers. To make that claim, we would
have to understand what's causing the failures, and whether the thing
they're testing is something that's likely to matter.

This came from https://github.com/mdn/browser-compat-data/pull/5963.