Note Safari bug resolving CSS variables and url()
In certain circumstances, Safari will incorrectly resolve `url()` values
relative to the document instead of the stylesheet. It's a bit of an
edge case, but there's two WebKit bugs and a BCD issue about it (#9075),
so I figured that justified a note.

See for details about the bug itself and the fix: https://trac.webkit.org/changeset/267951/webkit

Fixes #9075