Change Safari support from null to true based on tests
This updates features where both Safaris had nulls, where
mdn-bcd-collector results from Safari 14 on macOS and on iOS 14 both
indicate support.

This is an appropriate category to update in bulk, since false positives
in the tests are quite uncommon, and as an extra safety the true values
will have to be revisited to determine the exact versions later.