fix unhandledrejection_event edge compatibility
The current data says it's available in <= 79.  this conflicts with caniuse (https://caniuse.com/?search=unhandledrejection) as well as not making much sense and also conflicting with my own testing - I verified it's available in Edge 88; and it's also been in Chrome for a while.  As 79 is the first Edge based on the same code base as Chrome, it seems far more likely that Edge >=79 is the correct range of Edge versions with support - or certainly more correct than the <=79 range.