Remove AuthenticationAssertion, ScopedCredential and ScopedCredentialInfo
These were added in https://github.com/mdn/browser-compat-data/pull/1742,
but none of these interfaces are in Chrome, Firefox or Safari, or appear
in Chromium, Gecko or WebKit source code. (There are matches in Gecko
and WebKit, but for internal things, not web-exposed interfaces.)

`AuthenticationAssertion` was renamed to `AuthenticatorAssertionResponse`
in https://github.com/w3c/webauthn/pull/397.

`ScopedCredential` was renamed to `PublicKeyCredential` in
https://github.com/w3c/webauthn/pull/432.

`ScopedCredentialInfo` was renamed to `AuthenticatorResponse` in
https://github.com/w3c/webauthn/pull/397.

All of the new names are already in BCD.