Remove the api.Window.OverconstrainedError entry (#9401)
There's already an api.OverconstrainedError entry with subfeatures. The
compat data for that either matches or is more precise than that being
removed here.

The reason this entry is here is probably because `OverconstrainedError`
was originally not defined using Web IDL but with some spec prose
decribing the interface object.