Fix descriptions and Safari versions for canvas Path2D overloads
The type was incorrectly described as Path. Rather than changing the
IDs, just lowercase them, as this is much more typical in BCD, and the
name of the argument is also path, although this is not observable.

Safari version archeology:

5 of these were around when isPointInPath/isPointInStroke were added and
were most likely shipped together with Path2D:
https://trac.webkit.org/browser/webkit/trunk/Source/WebCore/html/canvas/CanvasRenderingContext2D.idl?rev=165927

See however https://github.com/mdn/browser-compat-data/issues/9423 about
the challenge of distinguishing Safari 6.1 and 7. 7 is used because
that's the version used for Path2D itself.

drawFocusIfNeeded is different, it was added at WebKit trunk 538.35.0:
https://trac.webkit.org/changeset/168476/webkit
https://trac.webkit.org/browser/webkit/trunk/Source/WebCore/Configurations/Version.xcconfig?rev=168476

The Path2D overload was added at WebKit trunk 601.1.1:
https://trac.webkit.org/changeset/173652/webkit
https://trac.webkit.org/browser/webkit/trunk/Source/WebCore/Configurations/Version.xcconfig?rev=173652

Both are after Safari 7 was released, so that can't be right. The
likely versions are instead inferred from the WebKit trunk version.