Add entries for webkitOfflineAudioContext
In addition to not representing prefixed support, much of the existing
data claimed support that was earlier than it really was.

Chrome versions for webkitOfflineAudioContext and OfflineAudioContext were
found with two mdn-bcd-collector tests:
https://mdn-bcd-collector.appspot.com/tests/api/webkitOfflineAudioContext
https://mdn-bcd-collector.appspot.com/tests/api/OfflineAudioContext/OfflineAudioContext

Other Chromium data was mirrored.

Safari 14 does not support the unprefixed constructor, that should be
coming in Safari 14.1.

Support for webkitOfflineAudioContext was confirmed in iOS 7, and so
assumed to have been introduced in iOS 6, matching webkitAudioContext.