Mirror Web Audio API shipping in Firefox 25 to Android (was 26) (#9515)
The existing data is inconsistent, with AudioScheduledSourceNode and
already ScriptProcessorNode marked as shipping in Firefox Android 25.

What actually happened is hard to tell from release notes and bug:
https://developer.mozilla.org/en-US/docs/Mozilla/Firefox/Releases/25
https://developer.mozilla.org/en-US/docs/Mozilla/Firefox/Releases/26
https://bugzilla.mozilla.org/show_bug.cgi?id=885505

This is made consistent because when updating existing data and notes in
BaseAudioContext, one has to assume one or the other. The difference
between 25 and 26 no longer matters, so just say 25.