Make Android data for BaseAudioContext more consistent
The origin of "33" appears to be when Web Audio was unprefixed:
https://github.com/mdn/browser-compat-data/pull/1890

(It was actually unprefixed in Chromium 35.)

The data references at the time does not inspire confidence:
https://web.archive.org/web/20180421031716/https://developer.mozilla.org/en-US/docs/Web/API/BaseAudioContext#Browser_compatibility
https://web.archive.org/web/20180309013210/https://developer.mozilla.org/en-US/docs/Web/API/BaseAudioContext/createPeriodicWave#Browser_compatibility

Instead set support to "18" like most of the Web Audio API and mirror
that to Samsung Internet which appears to have been mirrored from this.
WebView Android is similar, but set it to ≤37 like most of the Web Audio
API. It might have been supported earlier, but it's hard to test for.

The important part of this is the data previously said a lot of this was
shipped in Chrome for Android 33, but that's inconsistent with all the
other Web Audio API data, which says 18.