Update when webkitCreateShadowRoot was removed in Chromium
These tests were used:
http://staging-dot-mdn-bcd-collector.appspot.com/tests/api/Element/createShadowRoot
http://staging-dot-mdn-bcd-collector.appspot.com/tests/api/Element/webkitCreateShadowRoot

Chrome 34, 35 and 36 were tested to confirm the versions and that the
two did indeed overlap in Chrome 35, it wasn't a straight rename.

The data was mirrored for other Chromium-based browsers, achieving the
real point of this, namely avoiding version_added==version_removed for
the prefixed entry for webview_android.

Part of https://github.com/mdn/browser-compat-data/pull/9610.