Mark bug w/ text-decoration-thickness fixed from Chromium 89
Chromium bug 1154537 [1] – which meant that the `text-decoration-thickness` property did not work unless either `text-underline-offset` was set to something other than `auto` or `text-decoration-color` was set to something other than `currentColor` – has been fixed as of Chromium 89.

This has been verified to fix the issue (using https://codepen.io/36degrees/pen/JjRKqYv as a test case) in:

- Chrome 89
- Edge 89
- Opera 75, which is based on Chromium 89 [2]

I have not been able to verify it as fixed in Chrome or WebView on Android, but based on my understanding of how Chrome works I’d expect the issue to be resolved there as well.

[1]: https://crbug.com/1154537
[2]: https://blogs.opera.com/desktop/2021/01/opera-75-developer/