Update -webkit-mask-attachment removal versions (#9667)
This was removed in WebKit trunk version 537.20:
https://trac.webkit.org/changeset/136080/webkit
https://trac.webkit.org/browser/webkit/trunk/Source/WebCore/Configurations/Version.xcconfig?rev=136080

The Safari versions can be determined with fairly high confidence based
on this. When this was added in WebView Android is a little bit more
uncertain, but doesn't really matter since this entry will now be
eligible for removal as an obsolete feature.

Part of https://github.com/mdn/browser-compat-data/pull/9610.