Remove old flags for CSS Grid
These were all enabled more than two years ago and can be removed:
https://github.com/mdn/browser-compat-data/blob/main/docs/data-guidelines.md#removal-of-irrelevant-flag-data

The only irregular edit here is for the repeat() features of
grid-template-columns and grid-template-rows. There, the Firefox for
Android data was out of sync with Firefox, and was made to match.

Initial changes produced using @vinyldarkscratch's excellent script:
https://github.com/vinyldarkscratch/browser-compat-data/tree/scripts/remove-redundant-flags