Update Chromium version for ping content attributes
Chrome 12 is inferred from WebKit trunk version 534.27.0 when the IDL
attribute was added:
https://github.com/mdn/browser-compat-data/pull/9470#issuecomment-817091562

That came after the content attribute, but that behavior was behind a
preference, so this could still be slightly wrong.

However, support for the network requests was confirmed in Chrome 15, so
at most this could be wrong in a range between Chrome 12-15.