Remove never-shipped Bluetooth APIs
The many all-No entries makes the compat tables look very messy:
https://developer.mozilla.org/en-US/docs/Web/API/BluetoothDevice#browser_compatibility
https://developer.mozilla.org/en-US/docs/Web/API/BluetoothAdvertisingData#browser_compatibility
https://developer.mozilla.org/en-US/docs/Web/API/BluetoothRemoteGATTService#browser_compatibility

The removed APIs were all confirmed to not be supported on Chrome 90 on
macOS using these generated tests:
https://mdn-bcd-collector.appspot.com/tests/api/BluetoothDevice
https://mdn-bcd-collector.appspot.com/tests/api/BluetoothAdvertisingData
https://mdn-bcd-collector.appspot.com/tests/api/BluetoothRemoteGATTService

Additionally, Chromium's IDL shows most of this isn't there:
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/renderer/modules/bluetooth/bluetooth_device.idl;drc=02d4b05485a75ab044b42658284ae1134da1ebb9

The exception is watchAdvertisements/watchingAdvertisements which are
behind a flag, but removed anyway rather than filling out the data.

The getIncludedServices entry was also an outlier, it's not actually in
Chromium so the existing compat data was wrong.