Fix broken spec URLs for SVG xml:lang and <cursor> element
There’s no #XMLLangAttribute anchor in the SVG2 spec, so this change
replaces that anchor with the correct SVG2 anchor, #LangSpaceAttrs.

Also, there’s no <cursor> element in SVG2 (it was dropped), and so
no #CursorElement anchor — and so this change replaces the SVG2 spec URL
with a URL for the SVG1.1 (where that #CursorElement anchor does exist).