Demix SVGTests to SVGAnimationElement and SVGGraphicsElement
Support for requiredExtensions and systemLanguage in WebKit goes back to
when SVG was first shipped:
https://trac.webkit.org/browser/webkit/tags/old/Safari-522.11.1b/WebCore/ksvg2/svg/SVGTests.idl
https://trac.webkit.org/browser/webkit/branches/old/Safari-3-branch/WebCore/ksvg2/svg/SVGTests.idl

Support for SVGTests was added in Firefox 12:
https://developer.mozilla.org/en-US/docs/Mozilla/Firefox/Releases/12#svg

WebView ≤37 is used to match https://github.com/mdn/browser-compat-data/pull/10653.

mdn-bcd-collector has tests:
http://mdn-bcd-collector.appspot.com/tests/api/SVGAnimationElement/requiredExtensions
http://mdn-bcd-collector.appspot.com/tests/api/SVGAnimationElement/systemLanguage
http://mdn-bcd-collector.appspot.com/tests/api/SVGGraphicsElement/requiredExtensions
http://mdn-bcd-collector.appspot.com/tests/api/SVGGraphicsElement/systemLanguage

These weren't used to confirm the data since it goes so far back, but
mdn-bcd-collector at least doesn't constradict any of the added data.