Demix AbstractWorker to Worker/SharedWorker/ServiceWorker.onerror
Tests:
https://staging-dot-mdn-bcd-collector.appspot.com/tests/api/Worker/onerror
https://staging-dot-mdn-bcd-collector.appspot.com/tests/api/SharedWorker/onerror
https://mdn-bcd-collector.appspot.com/tests/api/ServiceWorker/onerror (prototype test only)

Manually confirmed:
 - Worker.onerror and SharedWorker.onerror in Chrome 15
 - *.onerror in Edge 80
 - ServiceWorker.onerror in Firefox 44
 - *.onerror in Firefox Android 80
 - Worker.onerror and SharedWorker.onerror in Opera 12.16
 - ServiceWorker.onerror in Safari 11.1

Also, the onerror attribute was in WebKit source at the time of the
initial release of Worker and SharedWorker, respectively:
https://trac.webkit.org/browser/webkit/releases/Apple/Safari%204.0/WebCore/workers/Worker.idl?rev=135093
https://trac.webkit.org/browser/webkit/releases/Apple/Safari%205.0/WebCore/workers/AbstractWorker.idl?rev=135093

From this we can also glean that AbstractWorker was at one point an
abstract base interface and not a mixin, but that doesn't matter.

Beyond this testing, much of the data is assumed to match the parent
entries, and the data ends up matching (modulo notes) for SharedWorker
and ServiceWorker. For Worker the only difference is the nodejs data,
which is set to false like onmessage and onmessageerror, rather than
working out the correct versions from the documentation:
https://nodejs.org/api/worker_threads.html#worker_threads_event_error