Update data for Service Worker APIs shipped in Safari 11.1 / iOS 11.3
The Safari 11.1 data comes from mdn-bcd-collector 3.1.4 results for
Safari 11 and 11.1. The iOS 11.3 data was confirmed by older collector
results for just iOS 11.3:
https://github.com/foolip/mdn-bcd-results/pull/1815

Other than the tests being wrong, the only room for error is that some
of this could have been shipped earlier in iOS, but in context that is
very unlikely and not worth verifying further.

Finally, the async_waitUntil is based on WebKit source, where the
current comment in ExtendableEvent::waitUntil makes it clear it's
supported, and the latest change to it was in trunk version 605.1.13
which predates the Safari 11.1 version:
https://trac.webkit.org/changeset/224584/webkit#file19
https://trac.webkit.org/browser/webkit/trunk/Source/WebCore/Configurations/Version.xcconfig?rev=224584