Sort out FetchEvent data
There was an unusual amount wrong here...

The current state of things in source:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/service_worker/fetch_event.idl;drc=fc05b522c6d4a85354c9d6f39b76c3d7f8448b68
https://github.com/mozilla/gecko-dev/blob/a4c86edb340aeee2b91a2ab617c66cb4ea7b7b66/dom/webidl/FetchEvent.webidl
https://github.com/WebKit/WebKit/blob/af262cabe3e5b067a177fb4203a7112f14028092/Source/WebCore/workers/service/FetchEvent.idl

The client attribute is removed as irrelevant as it actually never
shipped in Chrome, it was landed and reverted in 2015 and didn't reach
stable, and was removed in Gecko in 2015 before FetchEvent shipped:
https://chromium.googlesource.com/chromium/src/+/4e79b14a361c3960c11a2139df74a315b915c740
https://chromium.googlesource.com/chromium/src/+/ade5cc571923215463317b3212f7a50683f79038
https://github.com/mozilla/gecko-dev/commit/7029e804f87ef631b5307756f5b0f21c8d6c8b92

The navigationPreload attribute was the original name for
preloadResponse, but was renamed in Chromium before it reached stable:
https://chromium.googlesource.com/chromium/src/+/111601216214188fc2ace390da5b2af0095564fd
https://chromium.googlesource.com/chromium/src/+/95d26e23d47ff912eefe1d1caa7dc67552b3ffae

The preloadResponse attribute isn't in Gecko/WebKit source, see above.

The replacesClientId attribute turns out to not be in the source of any
engine (see above) and test shows it actually wasn't in Edge 18 either:
https://mdn-bcd-collector.appspot.com/tests/api/FetchEvent/replacesClientId

As a minor error, the resultingClientId attribute also wasn't in Edge 18:
https://mdn-bcd-collector.appspot.com/tests/api/FetchEvent/resultingClientId

Finally, the targetClientId was shipped in Safari, but in the spec it
was renamed to replacesClientId (see above) which hasn't shipped in any
browser: https://github.com/w3c/ServiceWorker/pull/1333

Nevertheless, treat Safari's targetClientId as an alternative name for
replacesClientId to combine the entries.