Remove entries representing indexed property getters (#9464)
These are curious, because the ability to access the members of these
Array-like interfaces is the whole point of them. There are many
interfaces like this, but only these few have entries for [].

These pages exist and will need to be removed:
https://developer.mozilla.org/docs/Web/API/DataTransferItemList/DataTransferItem
https://developer.mozilla.org/docs/Web/API/SourceBufferList/SourceBuffer
https://developer.mozilla.org/docs/Web/API/TrackDefaultList/TrackDefault

These pages do not exist:
https://developer.mozilla.org/docs/Web/API/CSSNumericArray/CSSNumericValue
https://developer.mozilla.org/docs/Web/API/CSSUnparsedValue/CSSUnparsedSegment

The interface pages do exist and show that support is largely identical
to the parent feature:
https://developer.mozilla.org/docs/Web/API/CSSNumericArray#Browser_compatibility
https://developer.mozilla.org/docs/Web/API/CSSUnparsedValue#Browser_compatibility
https://developer.mozilla.org/docs/Web/API/DataTransferItemList#Browser_compatibility
https://developer.mozilla.org/docs/Web/API/SourceBufferList#Browser_compatibility
https://developer.mozilla.org/docs/Web/API/TrackDefaultList#Browser_compatibility

The exception is SourceBufferList, but that must be an error in the data.