Refactor index.js to allow for better error messages
In order to print a useful error message, we need to know both which
file is being loaded, and the JSON path at which the problem occurred.

As a not-strictly-needed improvement, change how directories are
traversed to only create a single result object and extend that with
each file in order, avoiding intermediate result objects in the previous
recursive load() calls.

Fixes https://github.com/mdn/browser-compat-data/issues/10734.