Add Chrome/Safari versions for api.Element.requestPointerLock (#11164)
This PR adds real values for Chrome and Safari for the `requestPointerLock` member of the `Element` API.  By running `document.forms[0].webkitRequestPointerLock;` in various versions of Chrome and Safari, I have determined that Safari has never supported pointer locks before Safari 10.1, with or without a prefix.  Additionally, Chrome has removed the prefixed version of `requestPointerLock` in the following version after adding the non-prefixed version, so it is not present in Edge.