Update support for <menu>, <menuitem> and related APIs/events
Most of this was only ever implemented in Gecko, and is now behind the
"dom.menuitem.enabled" flag.

A few slices of it remain in HTML and should be supported, namely <menu>
itself, the HTMLMenuElement interface, and the compact attribute. The
latter however is obsolete, so marked as deprecated.

Existing data and notes were preserved as far as possible, even where
this leaves things in an inconsistent and likely incorrect state. This
is because all of this will probably eventually be removed and it would
be a lot of work to determine when exactly each facet of this was
supported in Firefox on different platforms. Assume by default that it
was from 8-85 without a flag, and behind a flag from 85.

Fixes https://github.com/mdn/browser-compat-data/issues/8947.