Rename and complete the entries for focus({preventScroll}) parameters
The naming now matches the guideline:
https://github.com/mdn/browser-compat-data/blob/main/docs/data-guidelines.md#parameters-and-parameter-object-features

Chromium support for this parameter for SVGElement came when the focus()
methods of HTMLElement and SVGElement were consolidated into
HTMLOrForeignElement. Before that only the HTMLElement variant had the
FocusOptions dictionary as an argument:
https://chromium.googlesource.com/chromium/src/+/5f506297541b9c0a041701401eeb133a16619bf9
https://storage.googleapis.com/chromium-find-releases-static/5f5.html#5f506297541b9c0a041701401eeb133a16619bf9

The other data is mirrored and IE is just assumed to be false.